+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.31.0
+
+## Component Enhancements
+
+* `Form`: Two new props have been added, `saveTheme` and `cancelTheme`, to apply available themes to Form buttons.
+
 # 0.30.0
 
 ## SettingsRow Component